requirements.txt
flask>=2.2.0
python-dotenv>=0.21.0
authlib>=1.0
requests>=2.28.1
flask_oauthlib>=0.9.6
